Move utils out to src root
