--- night_light/utils/__init__.py
@@ -1,1 +1,0 @@
-from .util_duckdb import *
